Change offline key configuration flow
Previously, offline key configuration allowed to add and remove keys in
parallel, which allowed correcting an addition/removal mistake right
away. The threshold was set in the end within the constraints of the
keystore.

This commit adopts a sequential flow of:
1. configure threshold
2. remove keys (ignoring threshold)
3. add keys (until user exit AND threshold is met)

Reasons for change:
- consistency with existing cli
- consistency with rest of dialog
- also allows to correct mistakes

Other changes:
Adds a custom IntPrompt implementation, which only accepts positive
ints.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>