Change output order in admin2 sign
As requested by Martin, this commit changes the output order to show
the root metadata to be signed prior to showing signature
requirements, and not ask the user about it.

The metadata is only showed once.

This commit also adds a test helper for expected output matching of the
pretty printed metadata.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>